+++ src/HealthSystem/HealthSystemKernel.java
@@ -1,3 +1,5 @@
+package HealthSystem;
+
 import components.standard.Standard;
 
 /**