+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2024.11.30
+
+### Added
+
+- Designed test suite for HealthSystem component
+- Designed two different use cases for HealthSystem component
+
 ## 2024.11.14
 
 ### Added